+++ backend/src/main/java/com/magicvs/backend/service/ScryfallMigrationService.java
@@ -6,7 +6,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.RestTemplate;
@@ -15,6 +14,8 @@ import tools.jackson.databind.JsonNode;
 import java.net.URI;
 import java.time.LocalDate;
 import java.util.*;
+
+import org.springframework.context.annotation.Profile;
 
 @Service
 public class ScryfallMigrationService {
@@ -44,27 +45,7 @@ public class ScryfallMigrationService {
     @Lazy
     private ScryfallMigrationService self;
 
-    /**
-     * Tarea programada diaria a las 3:00 AM para aplicar migraciones y luego sincronizar cartas Standard.
-     */
-    @Scheduled(cron = "0 0 3 * * ?")
-    public void runDailySyncAndMigrations() {
-        log.info("Iniciando sincronización diaria de cartas y aplicación de migraciones de Scryfall...");
-        try {
-            int migrationsCount = applyMigrations();
-            log.info("Se aplicaron {} migraciones de Scryfall.", migrationsCount);
-        } catch (Exception e) {
-            log.error("Error durante la aplicación programada de migraciones", e);
-        }
 
-        try {
-            log.info("Iniciando actualización de cartas Standard...");
-            int standardCount = scryfallService.importStandardCards();
-            log.info("Actualizadas/Importadas {} cartas Standard.", standardCount);
-        } catch (Exception e) {
-            log.error("Error durante la importación automática diaria de cartas Standard", e);
-        }
-    }
 
     /**
      * Consulta la API de Scryfall para obtener y aplicar las migraciones no procesadas.
